unificacion de controllers CompositeProject y Project, y DTOCompositeProject y DTOProject
