+++ src/main/java/com/project/service/CompositeProjectService.java
@@ -4,6 +4,7 @@ import com.project.DTO.DTOActionInsert;
 import com.project.DTO.DTOCompositeProject;
 import com.project.DTO.DTOCompositeProjectInsert;
 import com.project.DTO.DTOCompositeProjectUpdate;
+import com.project.DTO.DTOProject;
 import com.project.entities.CompositeProject;
 import com.project.entities.Entrepreneurship;
 import org.springframework.stereotype.Component;
@@ -13,23 +14,23 @@ import java.util.List;
 @Component
 public interface CompositeProjectService {
 
-    DTOCompositeProject postCompositeProject(DTOCompositeProjectInsert cp);
+    DTOProject postCompositeProject(DTOCompositeProjectInsert cp);
 
-    Iterable<DTOCompositeProject> getCompositeProjects();
+    Iterable<DTOProject> getCompositeProjects();
 
-    DTOCompositeProject getCompositeProject(Long id);
+    DTOProject getCompositeProject(Long id);
 
     CompositeProject getCompositeProjectEntity(Long id);
 
-    List<DTOCompositeProject> getCompositeProjectsThatContain(Long id);
+    List<DTOProject> getCompositeProjectsThatContain(Long id);
 
-    DTOCompositeProject addEntrepreneurship(Long main_project_id, Entrepreneurship e);
+    DTOProject addEntrepreneurship(Long main_project_id, Entrepreneurship e);
 
     boolean containsEntrepreneurship(Entrepreneurship mainProject,Entrepreneurship subProject);
 
     boolean containsCommonEntrepreneurships(Long main_project_id, Long subproject_id);
 
-    DTOCompositeProject updateCompositeProject(Long id, DTOCompositeProjectUpdate dto);
+    DTOProject updateCompositeProject(Long id, DTOCompositeProjectUpdate dto);
 
-    DTOCompositeProject postCompositeProjectAction(DTOActionInsert a, Long id);
+    DTOProject postCompositeProjectAction(DTOActionInsert a, Long id);
 }
